bouton ajouter a la creation que si on est connecter. Mis en place des accès par role et url. Gestion des permissions pour changer les roles mis en dur dans le code

+++ data/app/form/admin/role/deleteRole.php
@@ -1,14 +1,22 @@
 <?php
 
 if( isset($_GET['section']) and $_GET['section'] == 'delete-role' ){
-  $mysqli = bddConnect();
 
-  $query = "DELETE FROM `role` WHERE `role`.`id` = ".$_GET['id-role'];
-  bddQuery($mysqli, $query);
+  if ( isNumber($_GET['id-role']) ){
+    if( $_GET['id-role'] >= 1 && $_GET['id-role'] <= 4 ){
+      // ces ID ne sont pas autorisé a etre supprimés
+    }else{
+      $mysqli = bddConnect();
+      $query = "DELETE FROM `role` WHERE `role`.`id` = ".$_GET['id-role'];
+      bddQuery($mysqli, $query);
+      $mysqli->close();
+    }
+  }
+
   header('location: /admin/roles/');
-  exit('Role '.$_GET['id-user'].' delete<br />');
+  exit('Role '.$_GET['id-role'].' delete<br />');
 
-  $mysqli->close();
+
 }
 
  ?>
